Update cancel buttons and secondary buttons
