+++ src/components/buttons/index.js
@@ -1,3 +1,5 @@
 export { default as Button } from '~/components/buttons/Button';
 export { default as LinkButton } from '~/components/buttons/LinkButton';
 export { default as PrimaryButton } from '~/components/buttons/PrimaryButton';
+export { default as SecondaryButton } from '~/components/buttons/SecondaryButton';
+export { default as CancelButton } from '~/components/buttons/CancelButton';
