+++ src/dnsmanager/components/EditMXRecord.js
@@ -2,7 +2,8 @@ import React, { PropTypes, Component } from 'react';
 
 import { dnszones } from '~/api';
 import { ModalFormGroup } from '~/components/form';
-import { Form, Input, SubmitButton, CancelButton } from '~/components/form';
+import { Form, Input, SubmitButton } from '~/components/form';
+import { CancelButton } from '~/components/buttons';
 import { reduceErrors, ErrorSummary } from '~/errors';
 
 export default class EditMXRecord extends Component {
