+++ src/linodes/components/ConfigSelectModalBody.js
@@ -3,7 +3,7 @@ import React, { Component, PropTypes } from 'react';
 import { rebootLinode } from '~/api/linodes';
 import { hideModal } from '~/actions/modal';
 import { Button } from '~/components/buttons';
-import { CancelButton } from '~/components/form';
+import { CancelButton } from '~/components/buttons';
 
 export default class ConfigSelectModalBody extends Component {
   constructor(props) {
