+++ src/linodes/linode/networking/components/EditRDNS.js
@@ -1,7 +1,8 @@
 import React, { PropTypes, Component } from 'react';
 
 import { ModalFormGroup } from '~/components/form';
-import { Form, Input, SubmitButton, CancelButton } from '~/components/form';
+import { Form, Input, SubmitButton } from '~/components/form';
+import { CancelButton } from '~/components/buttons';
 import { reduceErrors, ErrorSummary } from '~/errors';
 import { setRDNS } from '~/api/linodes';
