+++ src/linodes/linode/settings/components/EditModal.js
@@ -2,8 +2,9 @@ import React, { Component, PropTypes } from 'react';
 
 import { linodes } from '~/api';
 import { resizeLinodeDisk } from '~/api/linodes';
-import { CancelButton, Form, FormGroup, FormGroupError,
+import { Form, FormGroup, FormGroupError,
   Input, SubmitButton } from '~/components/form';
+import { CancelButton } from '~/components/buttons';
 import { hideModal } from '~/actions/modal';
 import { ErrorSummary, reduceErrors } from '~/errors';
