+++ src/profile/integrations/components/CreateApplication.js
@@ -4,7 +4,8 @@ import { MAX_UPLOAD_SIZE_MB } from '~/constants';
 import { clients } from '~/api';
 import { updateClientThumbnail } from '~/api/clients';
 import { ModalFormGroup } from '~/components/form';
-import { Form, Input, SubmitButton, CancelButton } from '~/components/form';
+import { Form, Input, SubmitButton } from '~/components/form';
+import { CancelButton } from '~/components/buttons';
 import { reduceErrors, ErrorSummary } from '~/errors';
 import { renderSecret } from './CreatePersonalAccessToken';
