+++ src/profile/integrations/components/CreatePersonalAccessToken.js
@@ -4,7 +4,8 @@ import _ from 'lodash';
 import { ModalFormGroup } from '~/components/form';
 import { showModal, hideModal } from '~/actions/modal';
 import { ConfirmModalBody } from '~/components/modals';
-import { Form, Input, Select, SubmitButton, CancelButton } from '~/components/form';
+import { Form, Input, Select, SubmitButton } from '~/components/form';
+import { CancelButton } from '~/components/buttons';
 import { reduceErrors, ErrorSummary } from '~/errors';
 import { OAUTH_SUBSCOPES, OAUTH_SCOPES } from '~/constants';
 import { tokens } from '~/api';
